% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_exercise.R
\name{create_exercise}
\alias{create_exercise}
\title{Create an exercise}
\usage{
create_exercise(
  name,
  path = ".",
  lang = c("en", "nl"),
  descnames = list(),
  labels = character(0),
  boilerplate = FALSE
)
}
\arguments{
\item{name}{the name of the directory with the reading activity.}

\item{path}{the place to put this directory}

\item{lang}{the languages ("en" and or "nl") for english and dutch respectively}

\item{descnames}{a named list with optional descriptions for in the config.json file.}

\item{labels}{an optional vector with labels}

\item{boilerplate}{a logical value indicating whether you want a boilerplate, or a character value with the boilerplate code}
}
\value{
If the function finishes, the value \code{TRUE}.
}
\description{
This function creates the blueprint of an exercise,
including some boilerplate if needed. The files can then be
adapted manually later on. It will create a config.json file
specifically for R.
}
