% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reading.R
\name{create_reading}
\alias{create_reading}
\title{Create reading material}
\usage{
create_reading(
  name,
  path = ".",
  lang = c("en", "nl"),
  descnames = list(),
  labels = character(0)
)
}
\arguments{
\item{name}{the name of the directory with the reading activity.}

\item{path}{the place to put this directory}

\item{lang}{the languages ("en" and or "nl") for english and dutch respectively}

\item{descnames}{a named list with optional descriptions for in the config.json file.}

\item{labels}{an optional vector with labels}
}
\value{
TRUE if success, FALSE otherwise
}
\description{
With this function you can create a directory set up for a reading material.
}
